Update Protocol file structure
