+++ context/app/static/js/components/Detail/Protocol/index.js
@@ -1,0 +1,3 @@
+import Protocol from './Protocol';
+
+export default Protocol;
